fix: XYNE-17198 reviewers comments
Services-Requiring-Redeployment:
  - backend
  - dashboard
