+++ apps/dashboard/src/components/Chat/ChatDirectory/FilterChipNode.tsx
@@ -36,14 +36,14 @@ import { Hashtag, UserTwo, Lock02Close } from '@xyne/icons';
 import { ChannelScopeType, ChannelVisibility, TicketPriority } from '@xyne/shared';
 import { useChannel } from '../../../hooks/useChannels';
 import Avatar from '../../ui/Avatar/Avatar';
-import { MentionType, type MentionData } from './ChannelCommandMenu.types';
+import { ChipType, type ChipData } from './ChannelCommandMenu.types';
 
 // ---- Editable label (FilterChipNode) ----
 
 export type SerializedFilterChipNode = Spread<
   {
     type: 'filter-chip';
-    mentionData: MentionData;
+    mentionData: ChipData;
     expectedText: string;
   },
   SerializedTextNode
@@ -51,9 +51,9 @@ export type SerializedFilterChipNode = Spread<
 
 /** The pill's leading prefix (`from:`, `in:`, `priority:`), or null for a prefix-less
  *  mention chip. Rendered as its own node so the avatar/glyph can sit after it. */
-export function chipPrefixText(mentionData: MentionData): string | null {
-  // Priority is a value filter; its chip always reads `priority:` regardless of MentionData.
-  if (mentionData.type === MentionType.PRIORITY) {
+export function chipPrefixText(mentionData: ChipData): string | null {
+  // Priority is a value filter; its chip always reads `priority:` regardless of ChipData.
+  if (mentionData.type === ChipType.PRIORITY) {
     return 'priority:';
   }
   return mentionData.prefix ?? null;
@@ -62,31 +62,31 @@ export function chipPrefixText(mentionData: MentionData): string | null {
 /** The pill's value label — the only editable part of the chip. A prefix-less mention reads
  *  with its sigil (`@alice`/`#general`), not `from:`/`in:` (which would mislabel it as an
  *  author/scope filter). */
-export function chipLabelText(mentionData: MentionData): string {
+export function chipLabelText(mentionData: ChipData): string {
   // Chip always reads lowercase (`priority: high`), independent of the dropdown label.
-  if (mentionData.type === MentionType.PRIORITY) {
+  if (mentionData.type === ChipType.PRIORITY) {
     return mentionData.id.toLowerCase();
   }
   // Date chips read as the bare date; the prefix already says which edge it is.
-  if (mentionData.type === MentionType.DATE) {
+  if (mentionData.type === ChipType.DATE) {
     return mentionData.id;
   }
   // A mention filter reads the way a mention is written — `mentions: @alice` — because the
   // `@` is the thing being searched for, not a type marker. The avatar beside it doesn't
   // duplicate it the way the hash glyph would duplicate a `#`.
-  if (mentionData.prefix === 'mentions:' && mentionData.type === MentionType.USER) {
+  if (mentionData.prefix === 'mentions:' && mentionData.type === ChipType.USER) {
     return `@${mentionData.name}`;
   }
   if (mentionData.prefix) {
     return mentionData.name;
   }
-  const sigil = mentionData.type === MentionType.USER ? '@' : '#';
+  const sigil = mentionData.type === ChipType.USER ? '@' : '#';
   return `${sigil}${mentionData.name}`;
 }
 
 /** The pill's full text (`from: alice`) — what demotion leaves behind verbatim to re-arm the
  *  dropdown. Split across the prefix + label nodes, but always reads as one string. */
-export function buildChipText(mentionData: MentionData): string {
+export function buildChipText(mentionData: ChipData): string {
   const prefix = chipPrefixText(mentionData);
   const label = chipLabelText(mentionData);
   return prefix ? `${prefix} ${label}` : label;
@@ -94,7 +94,7 @@ export function buildChipText(mentionData: MentionData): string {
 
 export class FilterChipNode extends TextNode {
   /** The picked entity + filter metadata this chip represents. */
-  __mentionData: MentionData;
+  __mentionData: ChipData;
   /** Label text at creation time (value only — the prefix lives in its own node);
    *  `FilterChipPlugin` demotes the chip once it diverges. */
   __expectedText: string;
@@ -107,7 +107,7 @@ export class FilterChipNode extends TextNode {
     return new FilterChipNode(node.__mentionData, node.__text, node.__expectedText, node.__key);
   }
 
-  constructor(mentionData: MentionData, text?: string, expectedText?: string, key?: NodeKey) {
+  constructor(mentionData: ChipData, text?: string, expectedText?: string, key?: NodeKey) {
     const resolved = text ?? chipLabelText(mentionData);
     super(resolved, key);
     this.__mentionData = mentionData;
@@ -147,7 +147,7 @@ export class FilterChipNode extends TextNode {
     return false;
   }
 
-  getMentionData(): MentionData {
+  getMentionData(): ChipData {
     return this.__mentionData;
   }
 
@@ -156,7 +156,7 @@ export class FilterChipNode extends TextNode {
   }
 }
 
-export function $createFilterChipNode(mentionData: MentionData): FilterChipNode {
+export function $createFilterChipNode(mentionData: ChipData): FilterChipNode {
   return new FilterChipNode(mentionData);
 }
 
@@ -169,7 +169,7 @@ export function $isFilterChipNode(node: LexicalNode | null | undefined): node is
 export type SerializedFilterChipIconNode = Spread<
   {
     type: 'filter-chip-icon';
-    mentionData: MentionData;
+    mentionData: ChipData;
   },
   SerializedLexicalNode
 >;
@@ -193,7 +193,7 @@ export const PRIORITY_ICON_COLOR: Record<string, string> = {
 };
 
 /**
- * Resolves a channel chip's glyph by id (keeping MentionData lean): DM → person,
+ * Resolves a channel chip's glyph by id (keeping ChipData lean): DM → person,
  * public channel → hash, private → lock; hash while the channel isn't cached yet.
  * Mirrors ChannelIcon's privacy logic but uses a generic person for DMs, not an avatar.
  */
@@ -221,25 +221,25 @@ export function ChannelChipIcon({
   );
 }
 
-function ChipIcon({ mentionData }: { mentionData: MentionData }): React.JSX.Element {
+function ChipIcon({ mentionData }: { mentionData: ChipData }): React.JSX.Element {
   // Priority — checked first so it never falls through to the channel branch (which
   // calls `useChannel`). Glyph tinted by severity; pill stays blue.
-  if (mentionData.type === MentionType.PRIORITY) {
+  if (mentionData.type === ChipType.PRIORITY) {
     return <SignalHigh className={`${ICON_CLASS} ${PRIORITY_ICON_COLOR[mentionData.id] ?? ''}`} />;
   }
   // Date — like priority, a value filter with a glyph rather than an entity avatar.
-  if (mentionData.type === MentionType.DATE) {
+  if (mentionData.type === ChipType.DATE) {
     return <CalendarDays className={ICON_CLASS} />;
   }
   // Board — checked before the channel branch below, which would otherwise claim it and
   // call useChannel with a board id.
-  if (mentionData.type === MentionType.BOARD) {
+  if (mentionData.type === ChipType.BOARD) {
     return <LayoutGrid className={ICON_CLASS} />;
   }
   // User filters show the picked person's photo (the design's 16px avatar, rounded-sm = 4px);
-  // `Avatar` resolves the user from the id and falls back to initials, so MentionData stays
+  // `Avatar` resolves the user from the id and falls back to initials, so ChipData stays
   // lean. No presence dot — it reads as noise inside a 16px chip.
-  if (mentionData.type === MentionType.USER) {
+  if (mentionData.type === ChipType.USER) {
     return <Avatar userId={mentionData.id} size='xs' showActiveStatus={false} />;
   }
   // The channel lookup lives in its own component so `useChannel` is never called
@@ -250,10 +250,10 @@ function ChipIcon({ mentionData }: { mentionData: MentionData }): React.JSX.Elem
 /**
  * The chip's avatar/glyph, sitting between the prefix and the value label. A DecoratorNode
  * (not part of the label TextNode) so it can render a real React component; it carries the
- * label's MentionData. Contributes no text, so the pill's text stays `from: alice`.
+ * label's ChipData. Contributes no text, so the pill's text stays `from: alice`.
  */
 export class FilterChipIconNode extends DecoratorNode<React.JSX.Element> {
-  __mentionData: MentionData;
+  __mentionData: ChipData;
 
   static override getType(): string {
     return 'filter-chip-icon';
@@ -263,7 +263,7 @@ export class FilterChipIconNode extends DecoratorNode<React.JSX.Element> {
     return new FilterChipIconNode(node.__mentionData, node.__key);
   }
 
-  constructor(mentionData: MentionData, key?: NodeKey) {
+  constructor(mentionData: ChipData, key?: NodeKey) {
     super(key);
     this.__mentionData = mentionData;
   }
@@ -305,12 +305,12 @@ export class FilterChipIconNode extends DecoratorNode<React.JSX.Element> {
     return <ChipIcon mentionData={this.__mentionData} />;
   }
 
-  getMentionData(): MentionData {
+  getMentionData(): ChipData {
     return this.__mentionData;
   }
 }
 
-export function $createFilterChipIconNode(mentionData: MentionData): FilterChipIconNode {
+export function $createFilterChipIconNode(mentionData: ChipData): FilterChipIconNode {
   return new FilterChipIconNode(mentionData);
 }
 
@@ -481,14 +481,14 @@ export function $isFilterChipContainerNode(
  *  chips are the value alone (no prefix, no icon); a chip of the current user also gets
  *  Slack's self color. */
 export function $createFilterChip(
-  mentionData: MentionData,
+  mentionData: ChipData,
   currentUserId?: string,
 ): FilterChipContainerNode {
   const prefix = chipPrefixText(mentionData);
   // Any chip referencing the current user — bare @me or from:/with:/assignee: me — gets the
   // Slack self-mention color.
   const isSelfMention =
-    mentionData.type === MentionType.USER && !!currentUserId && mentionData.id === currentUserId;
+    mentionData.type === ChipType.USER && !!currentUserId && mentionData.id === currentUserId;
   const container = new FilterChipContainerNode(isSelfMention);
   if (prefix === null) {
     container.append($createFilterChipNode(mentionData));
@@ -514,7 +514,7 @@ export function $removeExistingPriorityChips(): void {
   const visit = (node: LexicalNode): void => {
     if ($isFilterChipContainerNode(node)) {
       const label = node.getChildren().find($isFilterChipNode);
-      if (label && label.getMentionData().type === MentionType.PRIORITY) {
+      if (label && label.getMentionData().type === ChipType.PRIORITY) {
         containers.push(node);
         return;
       }
