+++ apps/dashboard/src/routes/CallHistoryScreen/CallHistoryScreen.tsx
@@ -74,7 +74,7 @@ import { UpcomingCallsList } from '../../components/Call/UpcomingCallsList';
 import { useSearchMetrics } from '../../hooks/useSearchMetrics';
 import type { DisplaySearchResult } from '../../types/search';
 import { getUserDisplayName } from '../../utils/userDisplayName';
-import { MentionType, TabType } from '../../components/Chat/ChatDirectory/ChannelCommandMenu.types';
+import { ChipType, TabType } from '../../components/Chat/ChatDirectory/ChannelCommandMenu.types';
 import { type InitialQueryData } from '../../components/Chat/ChatDirectory/LexicalSearchInput';
 import { CallHistorySearchPanel } from './CallHistorySearchPanel';
 import { useCalendarSync } from '../../hooks/useCalendarSync';
@@ -305,7 +305,7 @@ const CallHistoryScreen = (): ReactElement => {
 
   const allUsers = useUsers();
   const allChannels = useAllChannels();
-  const [callMentionSearchType, setCallMentionSearchType] = useState<MentionType | null>(null);
+  const [callMentionSearchType, setCallMentionSearchType] = useState<ChipType | null>(null);
   const [callMentionSearchQuery, setCallMentionSearchQuery] = useState('');
   const [selectedMentionIndex, setSelectedMentionIndex] = useState(0);
   const [hasNavigatedMentions, setHasNavigatedMentions] = useState(false);
@@ -328,25 +328,25 @@ const CallHistoryScreen = (): ReactElement => {
   });
   const titleSearchQuery = searchQuery.trim();
   const userMentionResults = useActiveUserSearch(
-    callMentionSearchType === MentionType.USER ? callMentionSearchQuery : '',
+    callMentionSearchType === ChipType.USER ? callMentionSearchQuery : '',
     8,
   );
   const selectedCallSearchUserIds = useMemo(
     () =>
       callSearchSelectedMentions
-        .filter(mention => mention.type === MentionType.USER)
+        .filter(mention => mention.type === ChipType.USER)
         .map(mention => mention.id),
     [callSearchSelectedMentions],
   );
   const selectedCallSearchChannelIds = useMemo(
     () =>
       callSearchSelectedMentions
-        .filter(mention => mention.type === MentionType.CHANNEL)
+        .filter(mention => mention.type === ChipType.CHANNEL)
         .map(mention => mention.id),
     [callSearchSelectedMentions],
   );
   const channelMentionResults = useMemo(() => {
-    if (callMentionSearchType !== MentionType.CHANNEL) return [];
+    if (callMentionSearchType !== ChipType.CHANNEL) return [];
     const query = callMentionSearchQuery.trim().toLowerCase();
     const selected = new Set(selectedCallSearchChannelIds);
 
@@ -378,12 +378,12 @@ const CallHistoryScreen = (): ReactElement => {
     selectedCallSearchUserIds.length > 0 || selectedCallSearchChannelIds.length > 0;
   const callSearchInitialQuery = useMemo<InitialQueryData | null>(() => {
     const mentions = callSearchSelectedMentions
-      .filter(mention => mention.type === MentionType.USER || mention.type === MentionType.CHANNEL)
+      .filter(mention => mention.type === ChipType.USER || mention.type === ChipType.CHANNEL)
       .map(mention => ({
         id: mention.id,
         name: mention.name || mention.id,
         type: mention.type,
-        prefix: mention.type === MentionType.USER ? ('with:' as const) : ('in:' as const),
+        prefix: mention.type === ChipType.USER ? ('with:' as const) : ('in:' as const),
       }));
 
     return searchQuery || mentions.length > 0 ? { text: searchQuery, mentions } : null;
@@ -409,7 +409,7 @@ const CallHistoryScreen = (): ReactElement => {
         closeCallMentionSearch();
         return;
       }
-      setCallMentionSearchType(MentionType.USER);
+      setCallMentionSearchType(ChipType.USER);
       setCallMentionSearchQuery(query);
       setSelectedMentionIndex(0);
       setHasNavigatedMentions(false);
@@ -423,7 +423,7 @@ const CallHistoryScreen = (): ReactElement => {
         closeCallMentionSearch();
         return;
       }
-      setCallMentionSearchType(MentionType.CHANNEL);
+      setCallMentionSearchType(ChipType.CHANNEL);
       setCallMentionSearchQuery(query);
       setSelectedMentionIndex(0);
       setHasNavigatedMentions(false);
@@ -432,7 +432,7 @@ const CallHistoryScreen = (): ReactElement => {
   );
 
   const handleCallSearchChange = useCallback(
-    (text: string, mentions: Array<{ id: string; type: MentionType; prefix?: string }>) => {
+    (text: string, mentions: Array<{ id: string; type: ChipType; prefix?: string }>) => {
       if (isRestoringCallSearchRef.current) {
         if (!text && mentions.length === 0) return;
         isRestoringCallSearchRef.current = false;
@@ -441,19 +441,17 @@ const CallHistoryScreen = (): ReactElement => {
       setSearchQuery(text);
       setCallSearchSelectedMentions(
         mentions
-          .filter(
-            mention => mention.type === MentionType.USER || mention.type === MentionType.CHANNEL,
-          )
+          .filter(mention => mention.type === ChipType.USER || mention.type === ChipType.CHANNEL)
           .map(mention => {
             const existingMention = callSearchSelectedMentions.find(
               selected => selected.id === mention.id && selected.type === mention.type,
             );
             const user =
-              mention.type === MentionType.USER
+              mention.type === ChipType.USER
                 ? allUsers.find(candidate => candidate.id === mention.id)
                 : undefined;
             const channel =
-              mention.type === MentionType.CHANNEL
+              mention.type === ChipType.CHANNEL
                 ? allChannels.find(candidate => candidate.id === mention.id)
                 : undefined;
 
@@ -464,7 +462,7 @@ const CallHistoryScreen = (): ReactElement => {
                 (user ? getUserDisplayName(user) : channel?.name) ||
                 mention.id,
               type: mention.type,
-              prefix: mention.type === MentionType.USER ? 'with:' : 'in:',
+              prefix: mention.type === ChipType.USER ? 'with:' : 'in:',
             };
           }),
       );
