+++ apps/dashboard/src/routes/CallHistoryScreen/CallHistorySearchPanel.tsx
@@ -6,7 +6,7 @@ import { Switch } from '../../components/ui/Switch';
 import { Tooltip } from '../../components/ui/Tooltip/Tooltip';
 import { cn } from '../../utils/classNames';
 import { getUserDisplayName } from '../../utils/userDisplayName';
-import { MentionType } from '../../components/Chat/ChatDirectory/ChannelCommandMenu.types';
+import { ChipType } from '../../components/Chat/ChatDirectory/ChannelCommandMenu.types';
 import {
   LexicalSearchInput,
   type InitialQueryData,
@@ -21,11 +21,11 @@ interface CallHistorySearchPanelProps {
   syncMessage: CalendarSyncMessage | null;
   reauthCountdown: CalendarReauthCountdown | null;
   onCalendarSync: () => void;
-  callMentionSearchType: MentionType | null;
+  callMentionSearchType: ChipType | null;
   callMentionSearchQuery: string;
   callSearchSelectedMentions: Array<{
     id: string;
-    type: MentionType;
+    type: ChipType;
     prefix?: string;
     name?: string;
   }>;
@@ -42,7 +42,7 @@ interface CallHistorySearchPanelProps {
   closeCallMentionSearch: () => void;
   handleCallSearchChange: (
     text: string,
-    mentions: Array<{ id: string; type: MentionType; prefix?: string }>,
+    mentions: Array<{ id: string; type: ChipType; prefix?: string }>,
   ) => void;
   handleCallUserSearch: (query: string | null) => void;
   handleCallChannelSearch: (query: string | null) => void;
@@ -180,7 +180,7 @@ export function CallHistorySearchPanel({
               onEscapeKeyDown={closeCallMentionSearch}
               className='z-[9999] w-[var(--radix-popover-trigger-width)] overflow-hidden rounded-md border border-border bg-popover shadow-lg'
             >
-              {callMentionSearchType === MentionType.USER ? (
+              {callMentionSearchType === ChipType.USER ? (
                 filteredUserMentionResults.length > 0 ? (
                   <ul className='max-h-64 overflow-y-auto py-1'>
                     {filteredUserMentionResults.map((candidate, index) => (
@@ -223,7 +223,7 @@ export function CallHistorySearchPanel({
                 ) : (
                   <div className='px-3 py-3 text-sm text-muted-foreground'>No users found</div>
                 )
-              ) : callMentionSearchType === MentionType.CHANNEL ? (
+              ) : callMentionSearchType === ChipType.CHANNEL ? (
                 channelMentionResults.length > 0 ? (
                   <ul className='max-h-64 overflow-y-auto py-1'>
                     {channelMentionResults.map((channel, index) => (
